Fixed crash on shutdown after resizing
